Drop features, remove flask, use falcon
This drops every feature, removes flask, and switches to falcon.

Flask configuration is very opaque, and every additional flask module adds
more opaque configuration.  I'm not a fan of this, and is a large reason
why I chose to switch to Falcon, however for what I would like to see as
an end goal, flask is heavier than needed.  As well I was already
dropping a number of modules due to lack of Py3 support.

At this point, Python 3 is the only target.

gittle was switched to gitpython for py3 support.

The storage layer has been isolated from the web view, however it could
be improved.

The features that have been removed have been deleted from the web
templates.  Some JavaScript cleanup has occurred, however, the plan is
to remove the templates and JavaScript and switch to an Angular 2 app
written with TypeScript.

Falcon's default router does not like to glob paths, as such,
subdirectories are not working at this time, effectively limiting this
to a flat directory layout.  There isn't any plan to rectify this until
after app has been switched to the Angular 2 front end.